Connect mutate function
